feat: move dashboard edit button to app bar, visible only on dashboard route
